+++ samples/local-bigquery/1.import.sh
@@ -5,5 +5,6 @@ set -x
 gsutil rm -r gs://$GCP_BUCKET_NAME/tmp/*
 gsutil cp -r quickstart/ gs://$GCP_BUCKET_NAME/tmp/
 
+#SPARK_DRIVER_OPTIONS="-Dconfig.file=$PWD/application.conf -agentlib:jdwp=transport=dt_socket,server=y,suspend=y,address=5005 -Dlog4j.configuration=file://$SPARK_DIR/conf/log4j.properties.template"
 
 $SPARK_SUBMIT --driver-java-options "$SPARK_DRIVER_OPTIONS" $SPARK_CONF_OPTIONS  --class $COMET_MAIN $COMET_JAR_FULL_NAME import