+++ src/test/scala/ai/starlake/TestHelper.scala
@@ -94,7 +94,10 @@ trait TestHelper
         |COMET_METRICS_PATH="${cometTestRoot}/metrics/{domain}/{schema}"
         |COMET_AUDIT_PATH="${cometTestRoot}/audit"
         |COMET_UDFS="ai.starlake.udf.TestUdf"
-        |
+        |COMET_ACCESS_POLICIES_LOCATION="eu"
+        |COMET_ACCESS_POLICIES_TAXONOMY="RGPD"
+        |COMET_ACCESS_POLICIES_PROJECT_ID=${sys.env
+        .getOrElse("COMET_ACCESS_POLICIES_PROJECT_ID", "invalid_project")}
         |include required("application-test.conf")
         |""".stripMargin,
       ConfigParseOptions.defaults().setAllowMissing(false)
@@ -403,7 +406,7 @@ trait TestHelper
   // We need to start it manually because we need to access the HTTP mapped port
   // in the configuration below before any test get executed.
   lazy val kafkaContainer: KafkaContainer = {
-    val kafkaDockerTag = "7.0.0"
+    val kafkaDockerTag = "5.2.1"
     KafkaContainer.Def(kafkaDockerTag).start()
   }